+++ aftiktuna-rust/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 - New data file `character_profiles.json`, used for randomly-selected characters
 - New aftik colors and character profiles
+- Containers that can be searched for items (types include: tents)
 
 ### Changed
 
@@ -20,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Aftik corpses can now be set to use a color from a random character profile
 - Loot tables are no longer hardcoded, and can be defined in `assets/loot_table/`
 - Characters may now push crew members that are in the way for certain actions
+- The tent in the eyesaur forest is now a searchable container
 
 ### Fixed